Fixed Rotation (kind of)
The player ship can now rotate towards their direction of movement
regardless of the rotation of the camera. In a previous update, the
rotation of the ship wouldn't be correct if the camera's rotation
changed in any way.

--- PROBLEMS ---
- The player ship can't move entirely into the corners.
- The player ship has rotation applied that shouldn't be applied when
moving into the corners of the screen, when the player is pressing only
A/D or W/S.
- Different aspect ratios have access to different parts of the screen
due to clamping based on camera viewport size.