Changes: Movement and Rotation
Movement: Players can now move around with a controller, as long as that
controller is an Xbox controller, or can do XInput.

Rotation: Made adjustments to how far away the "LookAtPoint" is, and
found that the farther the point, the more space the player has to move
than before, at the cost of a less pronouced rotation which limits where
they can shoot.