Added Rotation and a Reticle
=== CHANGE LOG ===
- Players can now press "Q" and "E" on their keyboard (or LB and RB on
an xbox 360 controller) to rotate their ship to left and right.
- Rotating in a direction increases movement speed in that direction,
and slows movement speed in the opposite direction.
- A reticle has been added to help players aim their shots better.
However, this is an imperfect implementation of a reticle, as there are
times where it doesn't line up with the bullets.
- To prevent the reticle from disappearing behind objects in the level,
the player and reticle is now viewed by a different camera than the main
camera.